requirements.txt
httpx>=0.27
pydantic>=2.7
PyYAML>=6.0.1
jsonschema>=4.22.0
hypothesis>=6.98.0
schemathesis>=3.27.6
pytest>=8.2
jinja2>=3.1
uvicorn>=0.30
fastapi>=0.112
typer>=0.12
click>=8.1
colorama>=0.4
